requirements.txt
streamlit
bertopic
umap-learn
pandas
matplotlib
plotly
